tests/wpt/web-platform-tests/html/browsers/sandboxing/sandbox-inherited-from-required-csp.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Inherit sandbox from CSP embedded enforcement</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<body>
<script>
// Check sandbox flags are properly defined when its parent requires them and
// the child allows it.

const same_origin = get_host_info().HTTPS_ORIGIN;
const cross_origin = get_host_info().HTTPS_REMOTE_ORIGIN;
const check_sandbox_url =
  "/html/browsers/sandboxing/resources/check-sandbox-flags.html?pipe=";
const allow_csp_from_star = "|header(Allow-CSP-From,*)";

// Return a promise, resolving when |element| triggers |event_name| event.
const future = (element, event_name) => {
  return new Promise(resolve => {
    element.addEventListener(event_name, event => resolve(event))
  });
};

const check_sandbox_script = `
  try {
    document.domain = document.domain;
    parent.postMessage("document-domain-is-allowed", "*");
  } catch (exception) {
    parent.postMessage("document-domain-is-disallowed", "*");
  }
`;

const sandbox_policy = "sandbox allow-scripts allow-same-origin";

promise_test(async test => {
  const iframe = document.createElement("iframe");
  iframe.csp = sandbox_policy;

  // The <iframe> immediately hosts the initial empty document after being
  // appended into the DOM. It will, as long as its 'src' isn't loaded. That's
  // why a page do not load is being used.
  iframe.src = "/fetch/api/resources/infinite-slow-response.py";
  document.body.appendChild(iframe);

  const iframe_reply = future(window, "message");
  iframe.contentWindow.location =
    `javascript:${encodeURI(check_sandbox_script)}`;
  const result = await iframe_reply;
  iframe.remove();

  assert_equals(result.data, "document-domain-is-not-allowed");
}, "initial empty document");

promise_test(async test => {
  const iframe = document.createElement("iframe");
  iframe.src = "data:text/html,dummy";

  const iframe_load_1 = future(iframe, "load");
  document.body.appendChild(iframe);
  await iframe_load_1;

  const iframe_load_2 = future(iframe, "load");
  iframe.csp = sandbox_policy;
  iframe.src = "about:blank";
  await iframe_load_2;

  const iframe_reply = future(window, "message");
  iframe.contentWindow.location =
    `javascript:${encodeURI(check_sandbox_script)}`;
  const result = await iframe_reply;

  assert_equals(result.data, "document-domain-is-not-allowed");
}, "about:blank");

promise_test(async test => {
  const iframe = document.createElement("iframe");
  iframe.csp = sandbox_policy;
  iframe.src =
    `data:text/html,<script>${encodeURI(check_sandbox_script)}</scr`+`ipt>`;

  const iframe_reply = future(window, "message");
  document.body.appendChild(iframe);
  const result = await iframe_reply;

  assert_equals(result.data, "document-domain-is-disallowed");
}, "data-url");

promise_test(async test => {
  const iframe = document.createElement("iframe");
  iframe.csp = sandbox_policy;
  iframe.srcdoc = `<script>${check_sandbox_script}</scr`+`ipt>`;

  const iframe_reply = future(window, "message");
  document.body.appendChild(iframe);
  const result = await iframe_reply;

  assert_equals(result.data, "document-domain-is-not-allowed");
}, "srcdoc");

promise_test(async test => {
  const iframe = document.createElement("iframe");
  iframe.csp = sandbox_policy;

  const blob = new Blob(
    [ `<script>${check_sandbox_script}</scr`+`ipt>` ], { type: "text/html" });

  iframe.src = URL.createObjectURL(blob);

  const iframe_reply = future(window, "message");
  document.body.appendChild(iframe);
  const result = await iframe_reply;

  assert_equals(result.data, "document-domain-is-disallowed");
}, "blob URL");

promise_test(async test => {
  const iframe = document.createElement("iframe");
  iframe.csp = sandbox_policy;
  iframe.src = same_origin + check_sandbox_url + allow_csp_from_star;

  const iframe_reply = future(window, "message");
  document.body.appendChild(iframe);
  const result = await iframe_reply;

  assert_equals(result.data, "document-domain-is-disallowed");
}, "same-origin");

promise_test(async test => {
  const iframe = document.createElement("iframe");
  iframe.csp = sandbox_policy;
  iframe.src = cross_origin + check_sandbox_url + allow_csp_from_star;

  const iframe_reply = future(window, "message");
  document.body.appendChild(iframe);
  const result = await iframe_reply;

  assert_equals(result.data, "document-domain-is-disallowed");
}, "cross-origin");

</script>
