LayoutTests/imported/w3c/web-platform-tests/html/browsers/sandboxing/sandbox-inherited-from-required-csp-expected.txt



Harness Error (TIMEOUT), message = null

FAIL initial empty document assert_equals: expected "document-domain-is-disallowed" but got "document-domain-is-allowed"
FAIL about:blank assert_equals: expected "document-domain-is-disallowed" but got "document-domain-is-allowed"
PASS data-url
FAIL srcdoc assert_equals: expected "document-domain-is-disallowed" but got "document-domain-is-allowed"
FAIL blob URL assert_equals: expected "document-domain-is-disallowed" but got "document-domain-is-allowed"
FAIL same-origin assert_equals: expected "document-domain-is-disallowed" but got "document-domain-is-allowed"
FAIL cross-origin assert_equals: expected "document-domain-is-disallowed" but got "document-domain-is-allowed"
